requirements.txt
anthropic>=0.40.0
google-api-python-client>=2.100.0
youtube-transcript-api>=0.6.2
python-dotenv>=1.0.0
requests>=2.31.0
slack-bolt>=1.18.0
